LayoutTests/platform/win/fast/xmlhttprequest/xmlhttprequest-responsetype-arraybuffer-expected.txt

CONSOLE MESSAGE: line 59: TypeError: Type error
FAIL: Timed out waiting for notifyDone to be called
Tests XMLHttpRequest 'arraybuffer' loading with the .responseType and .response attributes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS 'arraybuffer' .response does not exist when .readyState is 1.
PASS responseType property exists.
PASS response property exists.
PASS xhr.responseType is initially set to default value of empty string.
PASS xhr.responseType has been correctly set to ''.
PASS xhr.responseType has been correctly set to 'text'.
PASS xhr.responseType has been correctly set to 'document'.
PASS exception correctly thrown when xhr.responseType is set to invalid value : Error: SYNTAX_ERR: DOM Exception 12.
PASS 'arraybuffer' .response does not exist when .readyState is 2.
PASS 'arraybuffer' .response does not exist when .readyState is 3.
FAIL 'arraybuffer' .response should exist when .readyState is 4.
PASS DONE LOADING
PASS received response object : null.
PASS exception correctly thrown when xhr.responseType is set to valid value too late in the loading process : Error: INVALID_STATE_ERR: DOM Exception 11.
